requirements.txt
nfl_data_py 
pandas 
numpy 
SQLAlchemy 
psycopg2-binary 
python-dotenv
requests
beautifulsoup4
lxml
matplotlib